Initial Add of Project
This project contains just the new HRCBootstrapWidgets.  It uses
CKEditor Framework 4.5.6 and supports the nesting of the bootstrap
widgets.
Nesting is a feature that was released with 4.5.5 but broken in 4.5.7
I have created this release to make sure that we have one that works.
We can not branch this project to add additional features and make sure
that we have a good one to fall back on.